refactor: rename *.js to *.jsx
