+++ src/components/App/App.jsx
@@ -1,9 +1,11 @@
+import React from 'react';
+import Game from '../Game/Game';
 import './App.css';
 
 function App() {
   return (
     <div className="App">
-
+      <Game />
     </div>
   );
 }
