+++ src/index.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import App from './components/App/App';
 import './index.css';
-import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById('root'),
 );
-
